Create separate class for PrivateObject
