+++ OptimizelySDK.Tests/OptimizelySDK.Tests.csproj
@@ -78,6 +78,7 @@
     <Compile Include="BucketerTest.cs" />
     <Compile Include="ProjectConfigTest.cs" />
     <Compile Include="UtilsTests\ConditionEvaluatorTest.cs" />
+    <Compile Include="UtilsTests\PrivateObject.cs" />
     <Compile Include="UtilsTests\ValidatorTest.cs" />
     <Compile Include="ValidEventDispatcher.cs" />
   </ItemGroup>
