+++ OptimizelySDK.Tests/OptimizelyTest.cs
@@ -24,7 +24,7 @@ using OptimizelySDK.Exceptions;
 using OptimizelySDK.Event;
 using OptimizelySDK.Entity;
 using NUnit.Framework;
-using System.Reflection;
+using OptimizelySDK.Tests.UtilsTests;
 
 namespace OptimizelySDK.Tests
 {
@@ -82,34 +82,7 @@ namespace OptimizelySDK.Tests
             Config = null;
             EventBuilderMock = null;
         }
-        private class PrivateObject
-        {
-            private object createdInstance;
-            private Type instanceType;
-
-            public PrivateObject(Type privateObject, Type[] parameterTypes, object[] parameterValues)
-            {
-                instanceType = privateObject;
-                createdInstance = Activator.CreateInstance(instanceType, parameterValues);
-            }
-            private PrivateObject(Type privateObject, object[] parameterValues)
-            {
-                instanceType = privateObject;
-                createdInstance = Activator.CreateInstance(instanceType, parameterValues);
-            }
-
-            public void SetFieldOrProperty(string propertyName, object value)
-            {
-                instanceType.InvokeMember(propertyName, BindingFlags.Instance | BindingFlags.Public | BindingFlags.SetProperty | BindingFlags.NonPublic | BindingFlags.SetField,
-                    Type.DefaultBinder, createdInstance, new object[] { value });
-            }
-
-            public object Invoke(string name, params object[] args)
-            {
-                return instanceType.InvokeMember(name, BindingFlags.Instance | BindingFlags.Public | BindingFlags.InvokeMethod | BindingFlags.NonPublic,
-                    Type.DefaultBinder, createdInstance, args);
-            }
-        }
+        
         private class OptimizelyHelper
         {
             static Type[] ParameterTypes = new[]
